pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my_assistant"
version = "0.1.0"
description = "A Streamlit‑based personal morning briefing agent (calendar, contacts, email)."
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name="Your Name", email="you@example.com" },
]
dependencies = [
  "streamlit",
  "google-auth",            # ← ensures google.oauth2 exists
  "google-auth-oauthlib",
  "google-api-python-client",
  "langchain",
  "langchain_community",
  "openai",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

